add like, bookmark, diskusi

+++ app/Models/User.php
@@ -48,25 +48,79 @@ class User extends Authenticatable
         ];
     }
 
+    /**
+     * Get the role of the user (admin, masyarakat, or pemerintah)
+     * @return string|null
+     */
+    public function getRoles()
+    {
+        $this->admin;
+        $this->masyarakat;
+        $this->pemerintah;
+    }
+
+    /**
+     * Relationship to Admin model
+     */
+    public function admin()
+    {
+        return $this->hasOne(Admin::class, 'id', 'id');
+    }
+
+    /**
+     * Relationship to Masyarakat model
+     */
+    public function masyarakat()
+    {
+        return $this->hasOne(Masyarakat::class, 'id', 'id');
+    }
+
+    /**
+     * Relationship to Pemerintah model
+     */
+    public function pemerintah()
+    {
+        return $this->hasOne(Pemerintah::class, 'id', 'id');
+    }
+
     // Relation to RatingReport
     public function likesReport()
     {
-        return $this->hasMany(RatingReport::class, 'id_user','id');
+        return $this->hasMany(RatingReport::class, 'id_user', 'id');
     }
 
-     // Check Like by report id
+    // Check Like by report id
     public function hasLikedReport($id_report)
     {
         return $this->likesReport()->where('id_report', $id_report)->exists();
     }
 
     // Toggle Like Report
-    public function toggleLikeReport($id_report){
+    public function toggleLikeReport($id_report, $loaded)
+    {
+        if ($loaded) {
+            $count = RatingReport::all()->where('id_report', $id_report);
+            return [
+                'count' => count($count),
+                'hasLikedReport' => $this->hasLikedReport($id_report),
+            ];
+        }
+
         $existingLike = $this->likesReport()->where('id_report', $id_report)->first();
         if ($existingLike) {
+            if ($loaded) {
+                $count = RatingReport::all()->where('id_report', $id_report);
+                return [
+                    'count' => count($count),
+                    'hasLikedReport' => true,
+                ];
+            }
+
             $existingLike->where('id_report', $id_report)->delete();
+            $count = RatingReport::all()->where('id_report', $id_report);
 
             return [
+                'count' => count($count),
                 'hasLikedReport' => false,
             ];
         } else {
@@ -74,8 +128,10 @@ class User extends Authenticatable
                 'id_report' => $id_report,
             ]);
         }
+        $count = RatingReport::all()->where('id_report', $id_report);
 
         return [
+            'count' => count($count),
             'hasLikedReport' => $this->hasLikedReport($id_report)
         ];
     }
@@ -83,7 +139,7 @@ class User extends Authenticatable
     // Relation to RatingBerita
     public function likesBerita()
     {
-        return $this->hasMany(RatingBerita::class, 'id_user','id');
+        return $this->hasMany(RatingBerita::class, 'id_user', 'id');
     }
 
     // Check Like by Berita id
@@ -93,13 +149,32 @@ class User extends Authenticatable
     }
 
     // Toggle Like Berita
-    public function toggleLikeBerita($id_berita){
+    public function toggleLikeBerita($id_berita,$loaded)
+    {
+        if ($loaded) {
+            $count = RatingBerita::all()->where('id_berita', $id_berita);
+            return [
+                'count' => count($count),
+                'hasLikedBerita' => $this->hasLikedBerita($id_berita),
+            ];
+        }
+
         $existingLike = $this->likesBerita()->where('id_berita', $id_berita)->first();
         if ($existingLike) {
+            if ($loaded) {
+                $count = RatingBerita::all()->where('id_berita', $id_berita);
+                return [
+                    'count' => count($count),
+                    'hasLikedBerita' => true,
+                ];
+            }
+
             $existingLike->where('id_berita', $id_berita)->delete();
+            $count = RatingBerita::all()->where('id_berita', $id_berita);
 
             return [
-                'toggleLikeBerita' => false
+                'count' => count($count),
+                'hasLikedBerita' => false
             ];
         } else {
             $this->likesBerita()->create([
@@ -107,8 +182,82 @@ class User extends Authenticatable
             ]);
         }
 
+        $count = RatingBerita::all()->where('id_berita', $id_berita);
         return [
+            'count' => count($count),
             'hasLikedBerita' => $this->hasLikedBerita($id_berita)
+        ];
+    }
+
+    // Relation to Diskusi
+    public function diskusi()
+    {
+        return $this->hasMany(Diskusi::class, 'id_user', 'id');
+    }
+
+    // Check Like by diskusi id
+    public function hasDiskusi($id_report)
+    {
+        return $this->diskusi()->where('id_report', $id_report)->exists();
+    }
+
+    // create disscusion
+    public function sendDiskusi($id_report, $content)
+    {
+        $this->diskusi()->create([
+            'id_report' => $id_report,
+            'content' => $content
+        ]);
+    }
+
+    // Relation to BookMark
+    public function bookmark()
+    {
+        return $this->hasMany(Bookmark::class, 'id_user', 'id');
+    }
+
+    // Check Like by report id
+    public function hasBookmark($id_report)
+    {
+        return $this->bookmark()->where('id_report', $id_report)->exists();
+    }
+
+    // toggle bookmark
+    public function toggleBookmark($id_report, $loaded)
+    {
+        if ($loaded) {
+            $count = Bookmark::all()->where('id_report', $id_report);
+            return [
+                'count' => count($count),
+                'hasBookmark' => $this->hasBookmark($id_report),
+            ];
+        }
+        
+        $existingBookmark = $this->bookmark()->where('id_report', $id_report)->first();
+        if ($existingBookmark) {
+            if ($loaded) {
+                $count = Bookmark::all()->where('id_report', $id_report);
+                return [
+                    'count' => count($count),
+                    'hasBookmark' => true,
+                ];
+            }
+            $existingBookmark->where('id_report', $id_report)->delete();
+            $count = Bookmark::all()->where('id_report', $id_report);
+            return [
+                'count' => count($count),
+                'hasBookmark' => false,
+            ];
+        } else {
+            $this->bookmark()->create([
+                'id_report' => $id_report,
+            ]);
+        }
+
+        $count = Bookmark::all()->where('id_report', $id_report);
+        return [
+            'count' => count($count),
+            'hasBookmark' => $this->hasBookmark($id_report)
         ];
     }
 }
